robo-juridico/requirements.txt
# ── Núcleo ────────────────────────────────────────────────────────────
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.2
python-dotenv>=1.0.0

# ── JWT auto-gerado via chave PEM (PDPJ / DJEN) ───────────────────────
PyJWT>=2.8.0
cryptography>=42.0.0

# ── Google Drive (opcional — descomente se usar SALVAR_NO_DRIVE=true) ─
# google-api-python-client>=2.100.0
# google-auth-httplib2>=0.2.0
# google-auth-oauthlib>=1.1.0
